Execute validation and quality control of GWAS summmary statistics

tidyGWAS() performs a set of validations on input colummns, repairs missing columns, and can add missing CHR/POS or RSID. In addition, CHR and POS is standardised to GRCh38, with coordinates on GRCh37 added in as well.

Briefly, tidyGWAS() updates RSID if possible using the refsnp-merged file from dbSNP. Each inputed column is then validated and coerced to the correct type.

If statistis such as P, B are missing, tidyGWAS() will attempt to impute them if possible using repair_stats()

Standard column names are assumed, BEFORE inputting into the function. This is a deliberate decision as automatic parsing of some important column names can be ambiguous For example, in some sumstats, A1 referes to effect allele, while other formats use A1 as non-effect allele.

Usage

tidyGWAS(
  tbl,
  dbsnp_path,
  ...,
  column_names,
  output_format = c("hivestyle", "parquet", "csv"),
  output_dir = tempfile(),
  CaseN = NULL,
  ControlN = NULL,
  N = NULL,
  build = c("NA", "37", "38"),
  default_build = c("37", "38"),
  indel_strategy = c("keep", "remove"),
  convert_p = 2.225074e-308,
  repair_cols = TRUE,
  logfile = FALSE
)

Arguments

tbl

a data.frame or character() vector

dbsnp_path

filepath to the dbSNP155 directory (untarred dbSNP155.tar)

...

pass additional arguments to arrow::read_delim_arrow(), if tbl is a filepath.

column_names

a named list of column names: list(RSID = "SNP", POS = "BP")

output_format

How should the finished cleaned file be saved?

'csv' corresponds to arrow::write_csv_arrow()

'parquet' corresponds to arrow::write_parquet()

'hivestyle' corresponds to arrow::write_dataset() split by CHR

output_dir

filepath to a folder where tidyGWAS output will be stored. The folder should not yet exist. Note that the default argument is tempfile(), meaning that tidyGWAS output will not be saved by default over R sessions.

CaseN

manually input number of cases

ControlN

manually input number of controls

N

manually input sample size

build

If you are sure of what genome build ('37' or '38'), can be used to skip infer_build() and speed up computation

default_build

If only RSID exists, the build cannot be inferred. Nonetheless, tidyGWAS applies a filter on incompatible alleles with GRCh37/38. In such a case, tidyGWAS needs to decide on which reference genome to compare alleles with.

indel_strategy

Should indels be kept or removed?

convert_p

What value should be used for when P-value has been rounded to 0?

repair_cols

Should any missing statistical columns be repaired if possible?

logfile

Should messages be redirected to a logfile?

Value

a dplyr::tibble()

Examples

if (FALSE) { # \dontrun{
tidyGWAS(tbl = "my_dataframe", logfile = "true", name = "test_run", outdir = "gwas_sumstat_dir")
} # }