Perform meta-analysis of GWAS summary statistics datasets cleaned by tidyGWAS

meta_analyze() will:

flip the effect allele to the reference allele (using either GRCh37 or GRCh38), control with ref

variant id is constructed using by columns

EAF (allele frequency) and INFO (imputation quality) are weighted by sample size, if present

CaseN, N, ControlN and EffectiveN are all summed and carried forward

Usage

meta_analyze(
  dset,
  by = c("CHR", "POS", "RSID", "EffectAllele", "OtherAllele"),
  method = c("ivw"),
  ref = c("REF_37", "REF_38")
)

Arguments

dset

an arrow::open_dataset() object

by

a character vector of column names to group by. Default is c("CHR", "POS", "RSID", "EffectAllele", "OtherAllele")

method

method to use for performing meta-analysis. Currently, only IVW (based on standard errors) is supported.

ref

either "REF_37" or "REF_38", depending on which column you want to use to standardize reference allele

Value

a dplyr::tibble()

Examples

if (FALSE) { # \dontrun{
dset <- arrow::open_dataset("path_to/sumstats/")
res <- meta_analyze(dset)
} # }