Execute validation and quality control of GWAS summmary statistics

`tidyGWAS()` performs a set of validations on input colummns, repairs missing columns, and can add missing CHR/POS or RSID. In addition, CHR and POS is standardised to GRCh38, with coordinates on GRCh37 added in as well.

Briefly, `tidyGWAS()` updates RSID if possible using the [refsnp-merged](https://ftp.ncbi.nih.gov/snp/latest_release/JSON/) file from dbSNP. Each inputed column is then validated and coerced to the correct type. Missing CHR/POS or RSID is detected and imputed using [repair_rsid()] or [repair_chr_pos()]. If both RSID and CHR:POS is present, [verify_chr_pos_rsid()] is executed to check that dbSNP CHR:POS:RSID agrees with CHR:POS:RSID in `tbl`.

If statistis such as `P`, `B` are missing, `tidyGWAS()` will attempt to impute them if possible using [repair_stats()]

Standard column names are assumed, BEFORE inputting into the function. This is a deliberate decision, as automatic parsing of some important column names can be ambigious For example, in some sumstats, A1 referes to effect allele, while other formats use A1 as non-effect allele. [tidyGWAS_columns()] can be used to standardise column names, and see the standard format.

Usage

tidyGWAS(
  tbl,
  ...,
  dbsnp_path,
  output_format = c("csv", "hivestyle", "parquet"),
  build = c("NA", "37", "38"),
  outdir = tempdir(),
  study_n,
  convert_p = 2.225074e-308,
  name = stringr::str_replace_all(date(), pattern = c(` ` = "_", `:` = "_")),
  keep_indels = TRUE,
  repair_cols = TRUE,
  logfile = FALSE,
  log_on_err = "tidyGWAS_logfile.txt",
  verbose = FALSE
)

Arguments

tbl

a `data.frame` or `character()` vector

...

pass additional arguments to [arrow::read_delim_arrow()] which is the function that will be called if tbl is a filepath possible arguments are `study_n` to set N, and `build` to set genome build.

dbsnp_path

filepath to the dbSNP155 directory (untarred dbSNP155.tar)

output_format

How should the finished cleaned file be saved? * "csv" corresponds to [arrow::write_csv_arrow()] * 'hivestyle' corresponds to [arrow::write_dataset()] split by CHR * 'parquet' corresponds to [arrow::write_parquet()]

build

Can be used to skip [infer_build()]

outdir

Where should results be saved after a succesful execution?

study_n

Use to set N in tbl

convert_p

What value should be used for P = 0?

name

name of the output directory

keep_indels

Should indels be kept?

repair_cols

Should any missing columns be repaired?

logfile

Should messages be redirected to a logfile?

log_on_err

Optional. Can pass a filepath to copy the logfile to when the function exists.

verbose

Explain filters in detail?

Value

a tibble or NULL, depending on outdir

Examples

if (FALSE) {
tidyGWAS(tbl = "my_dataframe", logfile = "true", name = "test_run", outdir = "gwas_sumstat_dir")
}